feat/ui
- created custom useTerminalSession hook
- seperated logic from main app.tsx
