+++ docker-compose.dev.yml
@@ -1,5 +1,3 @@
-version: '3.8'
-
 services:
   backend:
     build:
